Added new functionality
This new functionality consist in a button placed on the next piece view that, when the user deletes two rows, it is permited to change the piece that is incoming. The user can know that he can use the button when the led next to the word "next" is on green color.